+++ checker/src/main/java/org/checkerframework/checker/initialization/InitializationFieldAccessSubchecker.java
@@ -1,7 +1,10 @@
 package org.checkerframework.checker.initialization;
 
 import org.checkerframework.checker.compilermsgs.qual.CompilerMessageKey;
+import org.checkerframework.checker.signature.qual.FullyQualifiedName;
 import org.checkerframework.common.basetype.BaseTypeChecker;
+
+import java.util.List;
 
 /**
  * Part of the freedom-before-commitment type system.
@@ -21,6 +24,16 @@ public class InitializationFieldAccessSubchecker extends BaseTypeChecker {
 
     /** Default constructor for InitializationFieldAccessSubchecker. */
     public InitializationFieldAccessSubchecker() {}
+
+    /** Also handle {@code AnnotatedFor} annotations for the {@link InitializationChecker}. */
+    @Override
+    public List<@FullyQualifiedName String> getUpstreamCheckerNames() {
+        if (upstreamCheckerNames == null) {
+            super.getUpstreamCheckerNames();
+            upstreamCheckerNames.add(InitializationChecker.class.getName());
+        }
+        return upstreamCheckerNames;
+    }
 
     // Suppress all errors and warnings, since they are also reported by the InitializationChecker