+++ docs/CHANGELOG.md
@@ -3,6 +3,9 @@ Version 3.42.0-eisop5 (July ?, 2024)
 
 **User-visible changes:**
 
+Removed support for the `-Anocheckjdk` option, which was deprecated in version 3.1.1.
+Use `-ApermitMissingJdk` instead.
+
 **Implementation details:**
 
 **Closed issues:**