+++ docs/CHANGELOG.md
@@ -1,5 +1,5 @@
-Version 3.42.0-eisop5 (July ?, 2024)
-------------------------------------
+Version 3.42.0-eisop5 (November ??, 2024)
+-----------------------------------------
 
 **User-visible changes:**
 
@@ -16,7 +16,8 @@ Make `SourceChecker#suppressWarningsString` protected to allow adaptation in sub
 
 **Closed issues:**
 
-eisop#413, eisop#777, eisop#782.
+eisop#413, eisop#777, eisop#782, eisop#982.
+
 
 Version 3.42.0-eisop4 (July 12, 2024)
 -------------------------------------